Extract sourceJson.ts
