+++ src/main/apps/sourceChanges.ts
@@ -19,13 +19,12 @@ import { getAppsRootDir } from '../config';
 import describeError from '../describeError';
 import { addDownloadAppData } from './app';
 import { downloadAppInfos } from './apps';
+import { downloadSourceJson, writeSourceJson } from './sourceJson';
 import {
     addToSourceList,
-    downloadSourceJson,
     getSource,
     initialise,
     removeFromSourceList,
-    writeSourceJson,
 } from './sources';
 
 const downloadSource = async (
